+++ crates/bootc-migrate-core/src/lib.rs
@@ -41,6 +41,7 @@ pub mod mergetc;
 pub mod migration;
 pub mod motd;
 pub mod ostree;
+pub mod ostree_install;
 pub mod preflight;
 pub mod rebase_controller;
 pub mod rebase_plan;